Haipd: Pint all file

+++ tests/Feature/Controllers/LoanRepaymentControllerTest.php
@@ -26,20 +26,20 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $randomId = 2;
         $response = $this->postJson("api/loan-repayment/$randomId/pay", []);
         $response->assertStatus(Response::HTTP_UNAUTHORIZED);
-        $response = $this->withToken("Some random token")->postJson("api/loan-repayment/$randomId/pay", []);
+        $response = $this->withToken('Some random token')->postJson("api/loan-repayment/$randomId/pay", []);
         $response->assertStatus(Response::HTTP_UNAUTHORIZED);
     }
 
     public function test_validate_amount_format_on_pay()
     {
         $customerUser = $this->createCustomerUser();
-        $token = $customerUser->createToken("User Token");
+        $token = $customerUser->createToken('User Token');
         $loan = $this->generateLoan($customerUser->id);
         [$repayment] = $loan->repayments;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", []);
         $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
         $response->assertInvalid(['amount']);
-        $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => "Some random amount"]);
+        $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => 'Some random amount']);
         $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
         $response->assertInvalid(['amount']);
     }
@@ -50,7 +50,7 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $loan = $this->generateLoan($customerUser->id);
         $this->loanService->approveLoan($loan->id);
         $otherCustomer = $this->createCustomerUser();
-        $token = $otherCustomer->createToken("User token");
+        $token = $otherCustomer->createToken('User token');
         [$repayment] = $loan->repayments;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => $repayment->amount]);
         $response->assertStatus(Response::HTTP_FORBIDDEN);
@@ -61,7 +61,7 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
     {
         $customerUser = $this->createCustomerUser();
         $loan = $this->generateLoan($customerUser->id);
-        $token = $customerUser->createToken("User token");
+        $token = $customerUser->createToken('User token');
         [$repayment] = $loan->repayments;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => $repayment->amount]);
         $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
@@ -72,7 +72,7 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $customerUser = $this->createCustomerUser();
         $loan = $this->generateLoan($customerUser->id);
         $this->loanService->approveLoan($loan->id);
-        $token = $customerUser->createToken("User token");
+        $token = $customerUser->createToken('User token');
         [$repayment] = $loan->repayments;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => 10]);
         $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
@@ -83,7 +83,7 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $customerUser = $this->createCustomerUser();
         $loan = $this->generateLoan($customerUser->id);
         $this->loanService->approveLoan($loan->id);
-        $token = $customerUser->createToken("User token");
+        $token = $customerUser->createToken('User token');
         [$repayment] = $loan->repayments;
         $amount = $repayment->amount + 10;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => $amount]);
@@ -91,7 +91,7 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $response->assertJsonPath('id', $repayment->id);
         $response->assertJsonPath('status', LoanRepayment::STATUS_PAID);
         $response->assertJsonPath('paid_amount', function ($result) use ($amount) {
-            return (double)$result == (double)$amount;
+            return (float) $result == (float) $amount;
         });
         $this->assertDatabaseHas('loans', ['id' => $loan->id, 'status' => Loan::STATUS_PARTIAL_PAID]);
     }
@@ -101,14 +101,14 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $customerUser = $this->createCustomerUser();
         $loan = $this->generateLoan($customerUser->id);
         $this->loanService->approveLoan($loan->id);
-        $token = $customerUser->createToken("User token");
+        $token = $customerUser->createToken('User token');
         [$repayment] = $loan->repayments;
         $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => $repayment->amount]);
         $response->assertStatus(Response::HTTP_OK);
         $response->assertJsonPath('id', $repayment->id);
         $response->assertJsonPath('status', LoanRepayment::STATUS_PAID);
         $response->assertJsonPath('paid_amount', function ($amount) use ($repayment) {
-            return (double)$amount == (double)$repayment->amount;
+            return (float) $amount == (float) $repayment->amount;
         });
         $this->assertDatabaseHas('loans', ['id' => $loan->id, 'status' => Loan::STATUS_PARTIAL_PAID]);
     }
@@ -118,18 +118,17 @@ class LoanRepaymentControllerTest extends AbstractFeatureTest
         $customerUser = $this->createCustomerUser();
         $loan = $this->generateLoan($customerUser->id);
         $this->loanService->approveLoan($loan->id);
-        $token = $customerUser->createToken("User token");
+        $token = $customerUser->createToken('User token');
         $repayments = $loan->repayments;
         foreach ($repayments as $repayment) {
             $response = $this->withToken($token->plainTextToken)->postJson("api/loan-repayment/{$repayment->id}/pay", ['amount' => $repayment->amount]);
             $response->assertStatus(Response::HTTP_OK);
             $response->assertJsonPath('id', $repayment->id);
             $response->assertJsonPath('paid_amount', function ($amount) use ($repayment) {
-                return (double)$amount == (double)$repayment->amount;
+                return (float) $amount == (float) $repayment->amount;
             });
             $response->assertJsonPath('status', LoanRepayment::STATUS_PAID);
         }
         $this->assertDatabaseHas('loans', ['id' => $loan->id, 'status' => Loan::STATUS_PAID]);
     }
-
 }
